Refactor part2 countries
This commit adds a reject handler and modifies conditional rendering
logic to handle cases when fetching weather would fail because the
location query does not have any matches.

It also handles the edge case where typing a query like "georgia" would
show "Georgia" and "South Georgia" instead of the details for the
country "Georgia". Clicking on the `show` button will always show the
details for that country. Changing the query will either update the view
to show the matching countries (if less than 10 matches) or retain the
current matching country details.

Also made some reorganization and added comments for readability.